Deprecate use of rand::distributions::weighted module
