+++ src/distributions/mod.rs
@@ -100,8 +100,9 @@ pub use self::bernoulli::{Bernoulli, BernoulliError};
 pub use self::float::{Open01, OpenClosed01};
 pub use self::other::Alphanumeric;
 #[doc(inline)] pub use self::uniform::Uniform;
+
 #[cfg(feature = "alloc")]
-pub use self::weighted::{WeightedError, WeightedIndex};
+pub use self::weighted_index::{WeightedError, WeightedIndex};
 
 // The following are all deprecated after being moved to rand_distr
 #[allow(deprecated)]
@@ -155,7 +156,10 @@ pub mod uniform;
 #[cfg(feature = "std")] mod unit_circle;
 #[cfg(feature = "std")] mod unit_sphere;
 #[cfg(feature = "std")] mod weibull;
+
+#[deprecated(since = "0.8.0", note = "use rand::distributions::{WeightedIndex, WeightedError} instead")]
 #[cfg(feature = "alloc")] pub mod weighted;
+#[cfg(feature = "alloc")] mod weighted_index;
 
 mod float;
 #[doc(hidden)]
